Attributes Added ArrayTarget to be able to push data to attributes.
Drawing 4 triangles to screen using 2 different shaders using
a single vertex buffer with vertices data

Drawing 4 triangles to screen using multiple two vertex
buffers. One with vertices data and another with color data.
This is a non interleaved draw. Making sure that using a shader
with an attribute that is not used, doesn't accidentially
uses the data by disabling attributes after drawing.

Support for vertexbuffers in a non interleaved way.