Interleaved vertex buffers
To implement interleaved vertex buffer support I had to
split the old AttributeData interface into two parts
  1) AttributeData representing the actual data
  2) AttributePointer represeting a pointer into the data.

This means that the old VertexBuffer class now implements
both the AttributeData and AttributePointer intefaces.

The new InterleavedVertexBuffer is a composition of
  1) One AttributeData implementation called InterleavedData.
  2) One or more AttributePointer implementation called
     InterleavedPointer.

This way the arrayTarget can be used for both a plain
VertexBuffer and the individual parts of a InterleavedVertexBuffer.

Also extended the Resource interface to create a BufferResource
with additional method called create() that creates the buffer.