Indices and textures
Added support for drawing with indices.
Added support for textures.

Split the AttributeData interface into two parts.
1) BufferResource
2) Attribute pointer.

The buffer resource is just a buffer resource for data and the
attribute pointer is a pointer into some buffer resource.

The VertexBuffer implements both interfaces.

The interleaved vertex buffer is a composition of one BufferResource
implemented in a InterleavedData and one or more AttributePointers
implemented in InterleavedPointer.

IndicesBuffer is an implementation of a BufferResource.
There is a ElementsTarget in the backend context that can be
used to load, enable and disable indices when drawing.

The texture support is implemented by having a TextureUnitManager
in the backend resources. The TextureUnitManager has TextureUnits.
Each texture unit has a Texture2Dtarget (coming cube target soon).

Client can either use the default texture unit, or take/return
texture units from the TextureUnitManager.

Split the array target test into 3 pieces.