Render and frame buffer
Added framebuffer and renderbuffer support.

Moved up some methods in Texture2D class to TextureResource
interface to be able to re-use backend context code
for both a Texture2D (which has data, format, texeltype, width
and height) and an EmptyTexture which doesn't have any data
(but has format, texeltype, width and height).

Added a testcase that draws a triangle to a texture and then
draws the texture on a small quad on screen.

Added new interfaces for the various frame buffer attachments
like color, depth and stencil attachments.

A TextureResource is not a Resource, ColorAttachment and
DepthAttachment.

A Texture2D is a TextureResource.
A EmptyTexture is a TextureResource.
The backend context works with TextureResources.

The RenderBuffer is a a Resource, DepthAttachment and
StencilAttachment.