Handle action
Delegated the handling of the action to the source of the action.

The backend context is provided in the handling method.

Added a testcase where actions are handled in a generic way
by just looping over the geometries, getting the action
and letting the source handle them. After that the
action is removed.