Thread preparation
To ensure thread safety between the main and gl thread
the mutable parts are copied prior to starting the
rendering.

Each action in a ActionNotifier has a HandleThread
enum that decides which thread the action must be handled
on. The plan is to

1) drive updates on the main thread from the engine.

2) when the updates are finished the engine will collect
   all main thread actions and handle these on the main
   thread. This locks out the application main thread.

3) after point 2, all objects are immutable from a gl thread
   point of view (because it's only the engine on the main
   thread that handles the actions.

4) While the engine is rendering, the application is allowed
   to change objects on the main thread (because the rendering
   use copied data). If the application change objects,
   actions will be added and handled in the next frame.

Updated testcases with point 2 (above). Note that only the
Geometry, Material and Mesh are mutable. Backend resources
such as shaders, programs, indices buffer, textures, vertex buffers
are immutable. When using these primitives directly there is no need
to perform point 2 above.