GL actions
Added ActionNotifier to all resources. This way it's as simple to
handle GL actions in a unified way as it is for main thread actions.

Added testcases where actions needing gl context are handled (instead
of manually calling create/load for each resource.

Cleanup of testcases and naming in next commit.