Custom uniform
Added support of adding a custom uniform to a geometry.
The uniform can be added to multiple geometries using
same or different shaders.

When uniform is updated, notifications are sent to
all geometries that have the uniform which then makes
a copy of the data to ensure thread safety between
main Ui thread and gl thread.

When drawing the geometry, the uniform is included
in the draw method.