SceneGraph
[Definition : world dirty]

When a world get's dirty the following happens to all
nodes in the graph (depth first traversal)
  1) The Node world transform is marked.
  2) The Geometry engine uniforms are marked to be updated.

[New features]

1) Each node has a model transform.
   When the transform is changed for a node, all sub nodes are
   marked as world dirty.

2) Each node has a world transform.
   The world transform is an aggregation of all parents model transforms.
   The world transform is calculated on demand only.

3) The render node can have a camera.
   If the camera is updated the world gets dirty.
   If the camera frustrum is update the world gets dirty.

4) When a node is added to the graph the added node is marked
   with world dirty.

5) It's possible to search for a camera. This will traverse upwards
   until it find a camera.

6) It's possible to manually set the world dirty (Node method).

7) Added default engine uniforms. The following are supported.
    MODEL_MATRIX("u_modelMat", new Matrix44()),
    WORLD_MATRIX("u_worldMat", new Matrix44()),
    VIEW_MATRIX("u_viewMat", new Matrix44()),
    PROJECTION_MATRIX("u_projMat", new Matrix44()),
    WORLD_VIEW_MATRIX("u_worldViewMat", new Matrix44()),
    WORLD_VIEW_PROJECTION_MATRIX("u_worldViewProjMat", new Matrix44());

8) There is a engine uniform cache that can be querid given a shader
   program for active engine uniforms.

9) Added testcase to test drawing quad with a shader that uses
   the worldViewProjection matrix (mvp transform).

[Changed features]

1) ActionNotifier now requires that all actions are handled. This is
   done by checking boolean return value from handleAction method.
   This change was reuired in case of inheritance/overloading
   of the handleAction method (Node/Geometry). Actions that are
   not handled will result in exception.

[General]
Which ever class adds an action should implement the handleAction
method for the action. For example if the Node class adds an action
it's handleAction method should handle it. Since Geometry is a Node
as well, and faces that it also want to perform something at the time
the action is handled it should places it's own action as well. This
will keep the code cleaner and easier to understand.

Example : The node transform is changed. The Node will add an action
called NODE_WORLD_DIRTY. All actions can be sniffed (using the
onActionAdded method). The geometry will then add it's own action
called GEOMETRY_ENGINE_UNIFORM which means that engine uniforms
need to be invalidated. As state before, it would have been possible
for the geometry to act on the NODE_WORLD_DIRTY in the geometry
handleAction method but responsibility is weakend that way. It all
boils down to that objects that add actions should be responsible
for handling actions. Use the sniffer method (onActionAdded) to
create new actions instead.